renames context to [Name]Context to match the default exported type
