+++ src/signals/incident/components/form/ContainerSelect/ContainerList/ContainerList.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
 import ContainerList from './ContainerList';
 
-import { ContainerSelectProvider } from '../context';
+import { ContainerSelectProvider } from '../ContainerSelectContext';
 import { withAppContext } from 'test/utils';
 
 describe('signals/incident/components/form/ContainerSelect/ContainerList', () => {
