+++ src/signals/incident/components/form/ContainerSelect/ContainerSelectContext.test.tsx
@@ -4,9 +4,8 @@ import ReactDOM from 'react-dom';
 import { controls } from 'signals/incident/definitions/wizard-step-2-vulaan/afval';
 import { withAppContext } from 'test/utils';
 import type { ContainerSelectValue } from './types';
-import ContainerSelectContext, { ContainerSelectProvider } from './context';
+import ContainerSelectContext, { ContainerSelectProvider } from './ContainerSelectContext';
 import { render, screen } from '@testing-library/react';
-
 
 ReactDOM.createPortal = node => node as ReactPortal;
