+++ src/signals/incident/components/form/ContainerSelect/WfsLayer/WfsDataContext.tsx
@@ -7,7 +7,6 @@ export const NO_DATA: FeatureCollection = {
   features: [],
 };
 
-
 const initialValue: FeatureCollection = NO_DATA;
 
 const WfsDataContext = React.createContext(initialValue);
@@ -16,7 +15,8 @@ interface WfsDataProviderProps {
   value: FeatureCollection;
 }
 
-export const WfsDataProvider: FunctionComponent<WfsDataProviderProps> = ({ value, children }) =>
-  <WfsDataContext.Provider value={value}>{children}</WfsDataContext.Provider>;
-export default WfsDataContext;
+export const WfsDataProvider: FunctionComponent<WfsDataProviderProps> = ({ value, children }) => (
+  <WfsDataContext.Provider value={value}>{children}</WfsDataContext.Provider>
+);
 
+export default WfsDataContext;
